"Fix ambiguous increment"

+++ hilla_ride_ios/HelloTukTuk/Core/Services/RideRepository.swift
@@ -336,7 +336,7 @@ final class RideRepository {
                     "totalPlatformCommissionIqd": FieldValue.increment(Int64(split.platformCommissionIqd)),
                     "outstandingPlatformCommissionIqd": FieldValue.increment(Int64(split.platformCommissionIqd)),
                     "totalDriverEarningsIqd": FieldValue.increment(Int64(split.driverEarningsIqd)),
-                    "completedRidesCount": FieldValue.increment(1),
+                      "completedRidesCount": FieldValue.increment(Int64(1)),
                     "hasActiveRide": false
                 ], forDocument: driverRef)
             }
